backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
supabase==2.1.0
pinecone-client==2.2.4
deepface==0.0.79
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1